pyproject.toml
[tool.poetry]
name = "robot-bleu"
version = "0.1.0"
description = "Discord conversational bot using the ollama api"
authors = ["Alexandre Perdereau <alexandre.perdereau.01@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "robot_bleu"}]

[tool.poetry.dependencies]
python = "^3.10"
pyttsx3 = "^2.90"
discord-py = {extras = ["voice"], version = "^2.3.2"}
aiohttp = "^3.9.5"
elevenlabs = "^1.3.1"
transformers = "^4.41.2"
soundfile = "^0.12.1"
librosa = "^0.10.2.post1"
webrtcvad = "^2.0.10"
numpy = "^2.0.0"
scikit-learn = "^1.5.0"
torch = "^2.3.1"
python-dotenv = "^1.0.1"
ruff = "^0.4.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start-bot = "robot_bleu.main:main"